practica/requirements.txt
Django>=3.0,<4.0
sorl-thumbnail
Pillow==7.0.0
django-debug-toolbar==2.2
django-crispy-forms==1.9.0
django-messages-to-bootstrap-notify==0.2.6